+++ src/app/api/models/task-definition.ts
@@ -5,7 +5,9 @@ import {AppInjector} from 'src/app/app-injector';
 import {AlertService} from 'src/app/common/services/alert.service';
 import {DoubtfireConstants} from 'src/app/config/constants/doubtfire-constants';
 import {TaskDefinitionService} from '../services/task-definition.service';
-import {Grade, GroupSet, LearningOutcome, TutorialStream, Unit} from './doubtfire-model';
+import {Grade, GroupSet, LearningOutcome, Project, TutorialStream, Unit} from './doubtfire-model';
+import {Task} from './doubtfire-model';
+import {TaskPrerequisite} from './task-prerequisite';
 
 export type UploadRequirement = {
   key: string;
@@ -52,6 +54,9 @@ export class TaskDefinition extends Entity {
   similarityLanguage: string = 'c';
   hasJplagReport: boolean;
   assessInPortfolioOnly: boolean;
+
+  public readonly taskPrerequisitesCache: EntityCache<TaskPrerequisite> =
+    new EntityCache<TaskPrerequisite>();
 
   public readonly learningOutcomesCache: EntityCache<LearningOutcome> =
     new EntityCache<LearningOutcome>();
@@ -126,7 +131,10 @@ export class TaskDefinition extends Entity {
   public refresh(): void {
     const alerts = AppInjector.get(AlertService);
     AppInjector.get(TaskDefinitionService)
-      .fetch(this.id)
+      .fetch({
+        unitId: this.unit.id,
+        id: this.id,
+      })
       .subscribe({
         next: (taskDefinition) => {
           console.log(taskDefinition.name);
@@ -247,6 +255,12 @@ export class TaskDefinition extends Entity {
     }/scorm_data`;
   }
 
+  public get taskPrerequisiteUrl(): string {
+    return `${AppInjector.get(DoubtfireConstants).API_URL}/units/${this.unit.id}/task_definitions/${
+      this.id
+    }/prerequisites`;
+  }
+
   public get taskOverseerResourcesUploadUrl(): string {
     return `${AppInjector.get(DoubtfireConstants).API_URL}/units/${this.unit.id}/task_definitions/${
       this.id
@@ -285,5 +299,9 @@ export class TaskDefinition extends Entity {
     return httpClient
       .delete(this.taskOverseerResourcesUploadUrl)
       .pipe(tap(() => (this.hasTaskAssessmentResources = false)));
+  }
+
+  public projectTask(project?: Project): Task | undefined {
+    return project?.tasks?.find((p) => p.definition.id === this.id);
   }
 }