+++ src/app/api/models/unit.ts
@@ -28,7 +28,10 @@ import {
 } from './doubtfire-model';
 import {LearningOutcome} from './learning-outcome';
 import {AlertService} from 'src/app/common/services/alert.service';
-import { D2lAssessmentMapping } from './d2l/d2l_assessment_mapping';
+import {D2lAssessmentMapping} from './d2l/d2l_assessment_mapping';
+import {SidekiqJob} from './sidekiq-job';
+import {HttpClient} from '@angular/common/http';
+import {TaskPrerequisiteService} from '../services/task-prerequisite.service';
 
 export class Unit extends Entity {
   id: number;
@@ -576,17 +579,27 @@ export class Unit extends Entity {
     return `${AppInjector.get(DoubtfireConstants).API_URL}/csv/task_definitions?unit_id=${this.id}`;
   }
 
-  public downloadTaskCompletionCsv(): void {
-    AppInjector.get(FileDownloaderService).downloadFile(
-      `${AppInjector.get(DoubtfireConstants).API_URL}/csv/units/${this.id}/task_completion.json`,
-      `${this.name}-task-completion.csv`,
+  public downloadTaskCompletionCsv(): Observable<SidekiqJob> {
+    return AppInjector.get(HttpClient).get<SidekiqJob>(
+      `${AppInjector.get(DoubtfireConstants).API_URL}/csv/units/${this.id}/task_completion`,
     );
   }
 
-  public downloadTutorAssessmentCsv(): void {
-    AppInjector.get(FileDownloaderService).downloadFile(
-      `${AppInjector.get(DoubtfireConstants).API_URL}/csv/units/${this.id}/tutor_assessments.json`,
-      `${this.name}-tutor-assessments.csv`,
+  public downloadTasksAwaitingFeedbackCsv(): Observable<SidekiqJob> {
+    return AppInjector.get(HttpClient).get<SidekiqJob>(
+      `${AppInjector.get(DoubtfireConstants).API_URL}/csv/units/${this.id}/tasks_awaiting_feedback`,
+    );
+  }
+
+  public downloadTaskAssessmentCountsCsv(): Observable<SidekiqJob> {
+    return AppInjector.get(HttpClient).get<SidekiqJob>(
+      `${AppInjector.get(DoubtfireConstants).API_URL}/csv/units/${this.id}/task_assessment_counts`,
+    );
+  }
+
+  public downloadTutorAssessmentCsv(): Observable<SidekiqJob> {
+    return AppInjector.get(HttpClient).get<SidekiqJob>(
+      `${AppInjector.get(DoubtfireConstants).API_URL}/csv/units/${this.id}/tutor_assessments`,
     );
   }