+++ src/app/common/header/header.component.ts
@@ -9,6 +9,9 @@ import { AuthenticationService, Project, Task, Unit, UnitRole, User } from 'src/
 import { Subscription } from 'rxjs';
 import { MediaObserver } from 'ng-flex-layout';
 import { DoubtfireConstants, LogoSettings } from 'src/app/config/constants/doubtfire-constants';
+import {SidekiqJobEntry, SidekiqJobService} from 'src/app/api/services/sidekiq-job.service';
+import {SidekiqJobsModalService} from '../modals/sidekiq-jobs-modal/sidekiq-jobs-modal.service';
+import {QrModalService} from '../modals/qr-modal/qr-modal.service';
 
 @Component({
   selector: 'app-header',
@@ -38,6 +41,8 @@ export class HeaderComponent implements OnInit, OnDestroy {
   };
   private subscriptions: Subscription[] = [];
 
+  sidekiqJobs: SidekiqJobEntry[] = [];
+
   constructor(
     @Inject(calendarModal) private CalendarModal,
     @Inject(aboutDoubtfireModal) private AboutDoubtfireModal,
@@ -48,6 +53,9 @@ export class HeaderComponent implements OnInit, OnDestroy {
     private authService: AuthenticationService,
     protected media: MediaObserver,
     protected doubtfireConstants: DoubtfireConstants,
+    private sidekiqJobService: SidekiqJobService,
+    private sidekiqJobsModalService: SidekiqJobsModalService,
+    private qrModalService: QrModalService,
   ) {}
 
   ngOnInit(): void {
@@ -124,6 +132,33 @@ export class HeaderComponent implements OnInit, OnDestroy {
         }
       }),
     );
+
+    this.sidekiqJobService.sidekiqJobsSubject.subscribe((jobs) => {
+      this.sidekiqJobs = [...jobs];
+    });
+  }
+
+  showMyQr() {
+    const hostName = this.doubtfireConstants.API_URL.replace('/api', '');
+
+    const projectView =
+      this.currentProject &&
+      this.currentProject.student &&
+      this.currentProject.student.username !== this.currentUser.username;
+
+    const username = projectView ? this.currentProject.student.username : this.currentUser.username;
+
+    const url = `${hostName}/tutor-discussion?unitId=${this.currentUnit.id}&username=${username}`;
+    this.qrModalService.show(
+      url,
+      'Display this QR code during your class so your tutor can scan it to view your submissions and mark your tasks as complete.',
+      projectView ? `Viewing QR for ${this.currentProject.student.name}` : '',
+      true,
+    );
+  }
+
+  showSidekiqJob() {
+    this.sidekiqJobsModalService.show();
   }
 
   ngOnDestroy(): void {