+++ src/app/doubtfire-angular.module.ts
@@ -272,6 +272,13 @@ import {StaffNoteService} from './api/services/staff-note.service';
 import {LocalizedDatePipe} from './common/pipes/localized-date.pipe';
 import {StaffNotesViewComponent} from './projects/states/dashboard/directives/task-dashboard/directives/staff-notes-view/staff-notes-view.component';
 import {SubmissionTypeModalComponent} from './tasks/modals/submission-type-modal/submission-type-modal.component';
+import {SidekiqProgressModalComponent} from './common/modals/sidekiq-progress-modal/sidekiq-progress-modal.component';
+import {SidekiqJobsModalComponent} from './common/modals/sidekiq-jobs-modal/sidekiq-jobs-modal.component';
+import {SidekiqJobService} from './api/services/sidekiq-job.service';
+import {UnavailableCardComponent} from './errors/unavailable-card/unavailable-card.component';
+import {TaskDefinitionPrerequisitesComponent} from './units/states/edit/directives/unit-tasks-editor/task-definition-editor/task-definition-prerequisites/task-definition-prerequisites.component';
+import {TaskPrerequisitesCardComponent} from './projects/states/dashboard/directives/task-dashboard/directives/task-prerequisites-card/task-prerequisites-card.component';
+import {TaskPrerequisiteService} from './api/services/task-prerequisite.service';
 
 // See https://stackoverflow.com/questions/55721254/how-to-change-mat-datepicker-date-format-to-dd-mm-yyyy-in-simplest-way/58189036#58189036
 const MY_DATE_FORMAT = {
@@ -412,6 +419,11 @@ const MY_DATE_FORMAT = {
     QrModalComponent,
     LocalizedDatePipe,
     SubmissionTypeModalComponent,
+    SidekiqProgressModalComponent,
+    SidekiqJobsModalComponent,
+    UnavailableCardComponent,
+    TaskDefinitionPrerequisitesComponent,
+    TaskPrerequisitesCardComponent,
   ],
   // Services we provide
   providers: [
@@ -497,6 +509,8 @@ const MY_DATE_FORMAT = {
     FeedbackTemplateService,
     NestedCsvDownloadModalService,
     StaffNoteService,
+    SidekiqJobService,
+    TaskPrerequisiteService,
   ],
   imports: [
     FlexLayoutModule,