+++ src/app/projects/states/dashboard/directives/task-dashboard/directives/task-similarity-view/task-similarity-view.component.ts
@@ -75,12 +75,12 @@ export class TaskSimilarityViewComponent implements OnChanges {
       (_, response: HttpResponse<Blob>) => {
         // Open JPlag report viewer in embedded iframe
         setTimeout(() => {
-          this.jplagViewer.openReport(response.body);
+          this.jplagViewer.uploadReport(response.body);
           setTimeout(() => {
             // Open comparison between the two students
             this.jplagViewer.openComparison(
               similarity.task.project.student.username,
-              similarity.other_student.username,
+              similarity.otherStudent.username,
             );
             this.jplagOpenState = true;
           }, 100);